Survival & Inventory 1
The 4 consumables have been given functionality and resources can no longer be placed in the inventory hand slots. There are new pickups inheriting from InventoryItems which store data for a powerup, create a matching powerup and give it its data, and send the powerup to the player's PowerupManager when used so that the item's effect can be applied and/or removed when the duration ends.