S&I 8
Armor script has its own data for damage reduction and widening the range of survivable hot/cold temperatures. To work with this, the player has a new survival system called Defense which will be used before dealing damage to an object. The temperature system got a few additions to alter the temperatures the player takes damage from.