Tools & Weapons 2
Fixed spear animation so it won't start playing the second it spawns in. The player can now put the spear in a hand slot, the spear will appear to be held, child object to the player's camera, and disappear when changing to a different hand slot. Left click will play its animation. NOT FINISHED, THERE ARE NUMEROUS BUGS.