Tools & Weapons 3
There's now only a single item in the scene with unique name, this matches the item in the player's currently selected hand slot. The item is destroyed and instantiated whenever the player changes the selected slot, the slot items change, or the item is removed. The toggle that reverses hand slot selection using the scroll wheel now functions. Each item will store its own position relative to the player so any item can appear to be held when in the selected hand slot.