Tools & Weapons 4
Added a dummy with just the health and defense component. Spear confirmed does damage (had to add Rigidbodys in order for OnTriggerEnter to work) and it only does damage when the head of the spear connects during part of the animation. Giving the dummy more defense reduced the incoming damage. Stabbing also costs stamina, and if you don't have enough stamina it won't let you attack.