Tools & Weapons 5
The axe can do all the things the spear can do, with the addition of being used for a specific resource (which will have its own variation of health to check for this, enter TreeHealth) but will still be able to do a small amount of damage to enemies. TreeHealth also handles dropping wood with a set quantity at the tree's location when destroyed. Bringing stuff over from main.