Crafting & Recipes 2
Added a universal script to make recipes and add those recipes to the craft bench, needing the items required and the item it'll give. Added a rock and pickaxe prefab. Added a script that gives the craft bench its functionality regarding the UI and giving/taking items. Added ability for the player to open the craft bench UI using right click.