Gathering Material 7
Made an oven and oven UI prefab, gave the oven its own unique inventory (slight pain that it couldn't inherit from multiple classes) and UI slots for a fuel source, and item to convert, and the finished product. The oven's UI can be opened like the craft bench but items can't be dragged into the inventory slots yet.