Public Playtest Changes
Fixed some things related to the feedback we received. Created a copy of the forest scene so it isn't overridden later, this will be used for the build. There is now a menu opened with ESC which will pause the game and allow the player to resume or go to the main menu.