Crafting Bug 1
It hasn't been fixed but I finally know what is causing the bug. Track the debug logs to see for yourself. It all starts when the crafted item is placed in a new inventory slot. The item in that slot and the component from the recipe are shared, meaning when one changes, they both change. But I can't use the new keyword on my scripts, so I guess one option is making a new stick object, grabbing the component off of it, and deleting the stick object.