Expand Crafting 11
The crafting bug is fixed (cross your fingers). Hash codes in Debug Logs save lives. Essentially, the invItem component from the recipe and the invItem saved in the player's inventory would be the same, resulting in changes to the recipe changing the item in the player's inventory. Since Unity won't let me declare a new InvItem, I got around this by creating a new game object, getting the invItem component from that, changing it to match the recipe, and pass that component to the inventory.