Gameplay Loop 1
Bringing lots over from the Tank Game class, ways of handling audio, menus, settings, and data persistence without making a bunch of stuff from scratch. Code will need to be cut and merged until we only have what we need. Imported a sound to test audio levels, and it'll play and can be heard, but any attempts to change volume of the clip or within the mixer does nothing. It's not clear if the audio is even going through the mixer.