Spawn Points 3
Have confirmed spawn things work right. Now objects can spawn with a random +/- change to their x/z position, that radius is saved in the spawner. You can click the spawner with gizmos enabled to see how large of an area that will be (had to choose sphere, no option for just a circle). Additional things will for sure only spawn in if the spawn point is far enough away from the player or the player is somewhat far but the spawner is out of their FOV. Things will also spawn with a random rotation around the Y axis.