Loading Screen 1
Created a black screen that says "Loading" that appears when the forest scene is being loaded and goes away once the scene is fully loaded and the player isn't seeing things earlier than they should. Also added a credits button to the main menu that displays our names and what we're handling.