NavMesh 1
This is going to take more time to get perfect, but for the time being its good enough to do what we need. Everything in the scene is baked into the navmesh so any agents can path find around obstacles and anything that isn't static or is spawned in at runtime has the NavMeshObstacle component to carve out an area of the navmesh where it is in real time. Confirmed that the navmesh works by adding code to make the dummy move to an empty game object's position. The only issue I see right now is that AIs like the dummy would need to be obstacles for each other, but the obstacle component can't be added to an agent or else it completely screw over its pathfinding.