Tribesman AI 3
Added bool to Inventory Items to disable the ability for them to be picked up, fixing the duplication problem for items in hand that can be picked up during part of their use animation and the issue of stealing the Tribesman's spear. Confirmed, the Tribesman behavior is navigating between each state and is disabled when far enough away from the player.