Tribesman AI 4
Tried polishing Tribesman movement the best I can for moving towards the player or fleeing but I can't get it to move how I'd like it to with only NavMeshAgent.destination, and other attempts always end up with the agent not moving anywhere or going in the wrong direction, so it is what it is. At least the random movement plays nice.