Game Audio 1
Added all sound effects to the project and got most playing except for assets I still need to get from Marquez's branch. Also enable the audio listener on the main camera when the player dies and disable it when the player is spawned to get rid of any audio listener errors and let the player hear their own death sound.